backend/requirements.txt
Flask
Flask-PyMongo
Flask-CORS
Flask-JWT-Extended
passlib
python-dotenv
yfinance 